requirements.txt
pip~=23.3.1
psycopg2-binary~=2.9.9
pika~=1.3.2
flask~=3.0.0
python-dotenv~=1.0.0